Made two new channels txHallCallChan (master sends assignment to slave) and txRawHallCallChan (slave sends assignment to master to assign it to correct elevator).  In single_elevator.go these two channels are merged togheter such that the master only reads from the main hall call channel.
Modified SendHallAssignment to send a struct HallAssignmentMessage containing the targetElevator such that only the target elevator recives the order.

Modified RunOrderAssignment to only send hall call if it is not already assigned. RunSingleElevator also checks if the rawHallCallChan event is already registered before merging it with hallCallChan. ProcessButtonPress only forwards hallCalls to hallCallChan if they are not active.