Removed timer for obstruction in handler (that removed the elevator for active peers). It did not work.
Placed the timer in the fsm instead. in the case DoorOpen and Obstructed.

When the elevator is initialized the obstruction switch gets polled, and the elevator state is set to DoorOpen after moving to the inital floor. (Such that it stays in obstruction mode when it is restarted).

Moved init of hardware (config) above singleElevator initialization.